Pad to twice the size, FFT, then decimate
This is hacky and bad and I should feel bad.
Also there is some problem with simply decimating.  Probably requires a
change of scale factor or something.  Reconstructed image also requires
FFTshift to look anything like original object.  Possibly the phase is
getting messed up by the decimation because fs is not what it used to be.
Will require further investigation.