Re-vamp index limit calculations for consistency
Instead of calculating the index limits directly, first calculate a center
index, and then offset from that, rounding as necessary to get integers.
Rounding at each step ensures that the resulting sub-region always matches
the allotted size of the sub-images (m_s,n_s).

This is the same method used in Zheng's book.