Fix frequency limits (also fixes overlap problem)
Turns out that k is angular spatial frequency, so it's pi / pixel_size for
the limit of an image, or k0 * NA = 2pi/wavelength * NA for the objective.

The overlap mismatch was due to a discrepancy between the list of center
frequencies (kx_list,ky_list), the limits for the images (kt_max_rec), and
the objective bandwidth (kt_max_obj).

The oversampling mismatch was due to a missing factor of pi in kt_max_rec,
and a missing factor of 2 in kt_max_obj, resulting in the pi/2
discrepancy.