Rearrange to make work, and remove extra stuff
Some of this should have been removed back when the other extra stuff was
removed.  It's been broken since then, but this was never noticed because
variables remained in the workspace.  It is a good idea to clear the
workspace when testing.  For now, saving is disabled, will re-enable once
things actually work.