Use full complex object instead of only a quarter
pixel size must be half the size for the reconstruction, but if we do that
here, then 3/4 of the object does not get used.  Since we're just
generating the data, use all of it.  This can be seen by computing the
synthetic CTF and comparing it to the spectrum of the object.