Fix oversampling probelem once and for all
The previous few changes introduced a problem where the oversampled
sub-image regions would go outside the object's spectrum because it wasn't
oversampled at all.  This introduces just enough oversampling of the
object to prevent this, but still use as much of the complex object as
possible.

Many realizations happened during this commit, including the fact that the
pixel ratio of the large and sub-images shouldn't be the same as the
enhancement factor if their frequencies weren't related by the enhancement
factor.  Also that calculating pixel sizes first was convoluted and
unnecessary.  Half-bandwidths (kt_max_*) are calculated first, and pixel
sizes derived from those (do we even need pixel sizes?).

The biggest is that the object doesn't have to be oversampled by the same
percentage as the sub-images--just enough that they can still fit inside
it.  So the spatial frequency limit for the object is set to the synthetic
NA, plus a margin equal to the subimage oversampling.  The latter is only
the additional part, because the objective bandwidth is already contained
in the synthetic NA.  This is the reason for multiplying by the
oversampling factor - 1, instead of the raw factor.

On reflection, this might be equivalent to something like
k * NA_LED + k * NA_obj * oversampling_factor
but that's kind of irrelevant.