Use VideoDevice instead of videoinput for speed
There are two ways to acquire video data: using a videoinput object, and
using a VideoDevice system object.  Both are in the image acquisition
toolbox and appear to have the same functionality as far as we're
concerned.  The VideoDevice method seems to be lower-level or something.

Experimentation indicated acquisition to be much faster using the
VideoDevice method rather than the videoinput method, so we'll use that.

There are some differences to how these work, recommend reading the
documentation on VideoDevice objects:
https://www.mathworks.com/help/imaq/using-the-videodevice-system-object-to-acquire-frames.html
https://www.mathworks.com/help/imaq/imaq.videodevice.html

Major differences you need to worry about:

  - To setup object, use vid = imaq.VideoDevice(); instead of vid =
    videoInput()
      - VideoDevice also doesn't require an adaptor name, it can be called
        without arguments and will use the first adaptor
  - to acquire images, use image = step(vid); instead of
    image = getsnapshot(vid);
  - To set/get exposure or other parameters, use
    vid.DeviceProperties.<property>, rather than vidsrc.<property>.
      - This also means that vidsrc is not needed, and was removed
  - The data type of the image data can be set to double (or anything
    else), so there is no need to convert to double and scale to [0,1]
      - This is done with vid.ReturnedDataType = <type>;
  - There is no stoppreview() method, so you must use closepreview()