Correct indices
We've dealt with x,y vs m,n before, but never thought about i,j.  i should
correspond to y, and j to x.  It was wrong in both mock.m and
reconstruct.m, but also mock.m had them switched for the indices of
Images.  Apparently this is from 2ef6fab, and was a silly mistake at the
time.  It is likely the whole reason for the algorithm failing.